+++ brambl-sdk/src/main/scala/co/topl/brambl/display/package.scala
@@ -9,9 +9,14 @@ package object display
     with GroupDisplayOps
     with SeriesDisplayOps
     with TransactionDisplayOps
+    with QuivrDisplayOps
+    with ValidationErrorDisplayOps
     with BlockDisplayOps {
 
   val LabelLength = 27
+  val Indent = 2
+
+  def displayIndent(txt: String, indent: Int, prefix: String = " "): String = " " * indent + prefix + " " + txt
 
   def padLabel(label: String): String = {
     val padding = " " * (LabelLength - label.length).max(0)